Non-working implementation of XModem
Checking this in for comparison purposes but this does not implement Xmodem correctly because this can nack in the middle of packet transmission. I believe I need to check for <SOH> (vs. <EOT>) and if it is a packet, rip through the next 132 bytes and only then do my checking and send <NACK> or <ACK>.